Dev -> Master 0.2.1 sync

+++ MCLauncher/forms/Profile.cs
@@ -24,6 +24,8 @@ namespace MCLauncher.forms
         public static Profile Instance;
 
         public static bool isInitial;
+        public string origName;
+
 
         System.Windows.Forms.ToolTip toolTip = new System.Windows.Forms.ToolTip();
 
@@ -72,11 +74,13 @@ namespace MCLauncher.forms
 
             grbXboxGame.Text = Strings.grbGame;
             chkXboxDemo.Text = Strings.chkUseDemo.Substring(0, Strings.chkUseDemo.IndexOf("("));
+            lblXboxProfName.Text = Strings.lblProfName;
 
             tabControl1.TabPages[0].Text = Strings.cntHome;
             tabControl1.TabPages[1].Text = Strings.tabMods;
 
             profileName = profile;
+            origName = profileName;
             profMode = mode;
 
             listView1.Columns.Add(Strings.rowName);
@@ -87,6 +91,7 @@ namespace MCLauncher.forms
             {
                 editionBox.SelectedIndex = editionBox.Items.IndexOf("Java Edition");
                 nameBox.Text = profileName;
+                xboxNameBox.Text = profileName;
                 resXBox.Text = "854";
                 resYBox.Text = "480";
                 ramMaxBox.Value = 512;
@@ -103,10 +108,15 @@ namespace MCLauncher.forms
                 string manifest = Globals.client.DownloadString(Globals.javaManifest);
                 vj = JsonConvert.DeserializeObject<List<VersionManifest>>(manifest);
                 reloadVerBox("java");
+
+                deleteBtn.Visible = false;
+                openBtn.Visible = false;
+                saveBtn.Text = Strings.createProfile;
             }
             else if (profMode == "def")
             {
                 nameBox.Text = profileName;
+                xboxNameBox.Text = profileName;
                 resXBox.Text = "854";
                 resYBox.Text = "480";
                 ramMaxBox.Value = 512;
@@ -119,6 +129,7 @@ namespace MCLauncher.forms
 
                 version = dj.version;
                 nameBox.Text = profileName;
+                xboxNameBox.Text = profileName;
                 dirBox.Text = dj.directory;
                 string[] res = dj.resolution.Split(' ');
                 resXBox.Text = res[0];
@@ -195,9 +206,9 @@ namespace MCLauncher.forms
                 string manifest = "";
                 if (dj.edition == "java")
                     manifest = Globals.client.DownloadString(Globals.javaManifest);
-                else if(dj.edition == "x360")
+                else if (dj.edition == "x360")
                     manifest = Globals.client.DownloadString(Globals.x360Manifest);
-                else if(dj.edition == "javaedu")
+                else if (dj.edition == "javaedu")
                     manifest = Globals.client.DownloadString(Globals.javaEduManifest);
 
                 vj = JsonConvert.DeserializeObject<List<VersionManifest>>(manifest);
@@ -235,7 +246,7 @@ namespace MCLauncher.forms
             {
                 string[] row = { ver.type, ver.released.ToUniversalTime().ToString("dd.MM.yyyy HH:mm:ss") };
 
-                if(edition == "java")
+                if (edition == "java")
                 {
                     if (checkPreClassic.Checked && row[0] == "pre-classic")
                         listView1.Items.Add(ver.id + ver.alt).SubItems.AddRange(row);
@@ -273,29 +284,32 @@ namespace MCLauncher.forms
             listView1.Columns[1].Width = -1;
             listView1.Columns[2].Width = -2;
 
-            var item = listView1.FindItemWithText(version, true, 0, false);
+            if (listView1.Items.Count > 0)
+            {
+                var item = listView1.FindItemWithText(version, true, 0, false);
 
-            if (item != null)
-            {
-                listView1.Items[listView1.Items.IndexOf(item)].Selected = true;
-                listView1.EnsureVisible(listView1.Items.IndexOf(item));
-                saveBtn.Enabled = true;
-            }
-            else if (listView1.Items.Count == 0)
-                saveBtn.Enabled = false;
-            else
-            {
-                item = listView1.FindItemWithText(version, true, 0);
                 if (item != null)
                 {
-                    saveBtn.Enabled = true;
                     listView1.Items[listView1.Items.IndexOf(item)].Selected = true;
                     listView1.EnsureVisible(listView1.Items.IndexOf(item));
+                    saveBtn.Enabled = true;
                 }
+                else if (listView1.Items.Count == 0)
+                    saveBtn.Enabled = false;
                 else
                 {
-                    listView1.Items[0].Selected = true;
-                    listView1.EnsureVisible(listView1.Items[0].Index);
+                    item = listView1.FindItemWithText(version, true, 0);
+                    if (item != null)
+                    {
+                        saveBtn.Enabled = true;
+                        listView1.Items[listView1.Items.IndexOf(item)].Selected = true;
+                        listView1.EnsureVisible(listView1.Items.IndexOf(item));
+                    }
+                    else
+                    {
+                        listView1.Items[0].Selected = true;
+                        listView1.EnsureVisible(listView1.Items[0].Index);
+                    }
                 }
             }
         }
@@ -442,20 +456,44 @@ namespace MCLauncher.forms
                 }
             }
 
-            Directory.CreateDirectory($"{Globals.dataPath}\\instance\\{profileName}");
-            Directory.CreateDirectory($"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\");
-            if (!File.Exists($"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\mods.json"))
+            if (origName != profileName && profMode == "edit")
             {
-                File.WriteAllText($"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\mods.json", $"{{\"data\":1,\"items\":[]}}");
+                if (Directory.Exists($"{Globals.dataPath}\\instance\\{profileName}\\"))
+                {
+                    MessageBox.Show($"A profile with this name already exists, please use a different name.", "Profile", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                }
+                else
+                {
+                    Logger.Info("[Profile]", "Renaming the profile. This may take a while.");
+                    Directory.Move($"{Globals.dataPath}\\instance\\{origName}\\", $"{Globals.dataPath}\\instance\\{profileName}\\");
+                    File.WriteAllText($"{Globals.dataPath}\\instance\\{profileName}\\instance.json", saveData);
+
+                    string tempName = profileName; //loadInstanceList() overwrites profileName, so I had to do this shit lmao
+                    HomeScreen.loadInstanceList();
+                    HomeScreen.Instance.cmbInstaces.SelectedIndex = HomeScreen.Instance.cmbInstaces.FindString(tempName);
+                    HomeScreen.reloadInstance(tempName);
+
+                    this.Close();
+                }
             }
-            File.WriteAllText($"{Globals.dataPath}\\instance\\{profileName}\\instance.json", saveData);
+            else
+            {
+                Directory.CreateDirectory($"{Globals.dataPath}\\instance\\{profileName}");
+                Directory.CreateDirectory($"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\");
+                if (!File.Exists($"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\mods.json"))
+                {
+                    File.WriteAllText($"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\mods.json", $"{{\"data\":1,\"items\":[]}}");
+                }
 
-            string tempName = profileName; //loadInstanceList() overwrites profileName, so I had to do this shit lmao
-            HomeScreen.loadInstanceList();
-            HomeScreen.Instance.cmbInstaces.SelectedIndex = HomeScreen.Instance.cmbInstaces.FindString(tempName);
-            HomeScreen.reloadInstance(tempName);
+                File.WriteAllText($"{Globals.dataPath}\\instance\\{profileName}\\instance.json", saveData);
 
-            this.Close();
+                string tempName = profileName; //loadInstanceList() overwrites profileName, so I had to do this shit lmao
+                HomeScreen.loadInstanceList();
+                HomeScreen.Instance.cmbInstaces.SelectedIndex = HomeScreen.Instance.cmbInstaces.FindString(tempName);
+                HomeScreen.reloadInstance(tempName);
+
+                this.Close();
+            }
         }
 
         public bool nameCheck()
@@ -568,7 +606,7 @@ namespace MCLauncher.forms
             }
         }
 
-        public static void modListWorker(string mode, string name, string version, string file,  string type, string json, bool update)
+        public static void modListWorker(string mode, string name, string version, string file, string type, string json, bool update)
         {
             string indexPath = $"{Globals.dataPath}\\instance\\{profileName}\\jarmods\\mods.json";
             if (!File.Exists(indexPath))
@@ -834,7 +872,7 @@ namespace MCLauncher.forms
                 checkSnapshot.Visible = true;
                 checkExperimental.Visible = true;
 
-                if(!isInitial) //shitty fix but it doesn't crash anymore :troll:
+                if (!isInitial) //shitty fix but it doesn't crash anymore :troll:
                 {
                     string manifest = Globals.client.DownloadString(Globals.javaManifest);
                     vj = JsonConvert.DeserializeObject<List<VersionManifest>>(manifest);
@@ -887,6 +925,28 @@ namespace MCLauncher.forms
         private void btnOpenDotMc_Click(object sender, EventArgs e)
         {
             Process.Start($"{Globals.dataPath}\\instance\\{profileName}\\.minecraft\\");
+        }
+
+        private void deleteBtn_Click(object sender, EventArgs e)
+        {
+            DialogResult dialogResult = MessageBox.Show($"Are you sure you want to delete this profile?\nYou can't take this back!", "Profile", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialogResult == DialogResult.Yes)
+            {
+                Directory.Delete($"{Globals.dataPath}\\instance\\{profileName}", true);
+                HomeScreen.loadInstanceList();
+                HomeScreen.Instance.cmbInstaces.SelectedIndex = 0;
+                this.Close();
+            }
+        }
+
+        private void xboxNameBox_TextChanged(object sender, EventArgs e)
+        {
+            nameBox.Text = xboxNameBox.Text;
+        }
+
+        private void nameBox_TextChanged(object sender, EventArgs e)
+        {
+            xboxNameBox.Text = nameBox.Text;
         }
     }
